MIDebugJob: don't leak the debug session if config is invalid
The constructor of a MI debug job creates a debug session. When the
configured launch executable or its arguments are invalid, MIDebugJob
doesn't call MIDebugSession::startDebugging() and finishes with an
error. The debug session is not destroyed and the Debug area remains
active in KDevelop then.

Calling MIDebugSession::stopDebugger() finishes the debug session and
lets DebugController destroy it. This function explicitly and correctly
handles stopping debugger when it is not started.

BUG: 369313
FIXED-IN: 5.11.230400