MIExamineCoreJob: don't leak the debug session if dialog is rejected
The constructor of a MI debug job creates a debug session. When the
Select Core File dialog is rejected, MIExamineCoreJob doesn't call
MIDebugSession::examineCoreFile() and finishes. The debug session is not
destroyed and the Debug area remains active in KDevelop then.

Calling MIDebugSession::stopDebugger() finishes the debug session and
lets DebugController destroy it. This function explicitly and correctly
handles stopping debugger when it is not started.