WIP: Cache types in the clang DUChain builder
Introduce a TypeCacheKey that wraps a CXType and can be put
into a QHash. Because there's no clang_hashCursor equivalent for
CXType we hash the type spelling string instead.

Then, we cache the types we create in `makeType`. When we have a
hit, which happens pretty often esp. for trivial types, we can take
the type from the cache directly. Sadly, we still have to clone
the type, as the type modifiers will be added later and might be
different to what we see when `makeType` is called. But the clone
operation is still much faster than what `makeType` would do normally.
And especially for complex template types, we can safe a lot of time
with this cache. Previously, the worst-case scenario for the test
file from LLVM attached to bug 464198 took minutes to finish (I never
waited that long I have to admit). With this patch here, running
`duchainify -wuV mangle-ms-back-references-pr13207.cpp` finishes in
a mere 0.3s on my system - and this is with a debug build of kdevelop.

TODO: we might actually get away with a TU-global cache with the fixes
now, retest!

Note that the cache is currently _not_ shared across files in a TU,
because that could lead to bad types in situations that currently work
fine: Because we visit the AST of the TU multiple times, each time to
build the declarations for a single file, we can get into trouble for
complex situations with non-selfcontained files that rely on the
order of includes. Currently, this already leads to wrong types, but
only in the included files in isolation. If we would cache those types
though, all following files would also use the broken types. This
can be seen by setting `DEBUG_TYPE_CACHE=true` and running our tests
in test_duchain-clang already.

BUG: 464198
FIXED-IN: 5.11.230400