~MIDebugJobBase: don't access destroyed session
Since a recent commit "KCompoundJob: connect to subjob's finished, not
result signal", killing a KSequentialCompoundJob (and therefore its
derived class ExecuteCompositeJob) no longer kills unstarted subjobs.
Previously an unstarted debug job was killed, and thus finished, before
its session was destroyed. Now an unstarted debug job finishes in the
destructor when its session is already destroyed. The destructor then
attempts to stop the debugger of the destroyed session, which causes a
crash.

Note that not killing an unstarted debug job does not delay stopping
debugger on KDevelop exit, because the active session's debugger is
stopped in DebugController::cleanup() before the ExecuteCompositeJob
that contains a debug job is killed in RunController::cleanup().

Other regressions similar to this debug job crash are unlikely, because
a well-behaved job doesn't do much in constructor, and thus does not
perform risky cleanups in the destructor if unstarted. Ideally, a debug
job should create a debug session only after it is started.