Make the TestItemRepository tests determinisitc
Call srand with hardcoded seed values to ensure we can get the
same behavior when we run this test to simplify development.
A follow-up patch actually uncovers an assertion that only was
hit sometimes without this patch.

Also call srand from testLeaks to ensure we can get the same
behavior when we run all tests or just this one test on its own.