Cleanup: introduce central ItemRepository::allocateNextBuckets
For now this is mostly to centralize this logic, but in a follow-up
we will expand the logic in this function further.

Note that the ItemRepository::m_buckets member is still initialized
on-demand to 10 as before, either when `open` is called (which
happens through `ItemRepositoryRegistry::registerRepository`), or
when an `index` write operation is encountered the first time,
at which point `useBucket >= m_buckets.size()` will trigger a call
to `allocateNextBuckets`.