Don't try to insert items into existing monster buckets
When a bucket index is chosen for an item which falls into a
monster bucket, then there is no space to insert a new item.
Previously that was not checked at all, and could lead to
assertions below as uncovered by test_itemrepository in future
commits which also enable more of the debug facilities of the
ItemRepository code.