Cleanup: move m_freeSpaceBuckets lookup to usage
The lookup depends on the useBucket value which is set in various
places. Doing it once sounds dangerous, as it might be potentially
for a wrong useBucket value, when that changes later.

The assertion in the lower branch is removed altogether, as it
would otherwise fire. Follow-up patches are going to cleanup the
m_freeSpaceBuckets handling anyways, so it's OK to remove this here.
The current status is simply pretty damn broken...