ItemRepository: make reading from memory safer
Any char* pointer could be passed to readValue()'s char*& from argument
by accident. Now the pointer is passed once to FileMap's constructor,
and then the interface of the class is used. The new class also ensures
that the data pointer is not modified accidentally, but only when
something is read from it.