Use QRegularExpression for project filters
This requires some custom code to convert the old patterns using the
QRegExp::WildcardUnix format over to real regular expressions. Sadly,
there doesn't seem to be any existing code for that, so instead I
leveraged `QRegularExpression::wildcardToRegularExpression` as a start
and then changed it based on our local assumptions and expectations.
Most notably, we don't expect native paths, but we do want to allow
escape sequences in the patterns.