Allow project filters to supply patterns as regular expression
Use that then to improve the default pattern to exclude hidden
files/folders. Previously the pattern matched anything that had
a hidden folder anywhere in the path. Now we only match the last
segment of the path, meaning files in a hidden folder can be included
contrary to what we had before.

CCBUG: 466044