Add an option to store precompiled preambles in memory
If the temporary directory does not use tmpfs, there is the usual
tradeoff between RAM and disk storage.

Store preambles in memory by default as this has two extra advantages:
1. They are guaranteed to disappear on exit even if KDevelop crashes.
2. KDevelop's memory usage is more explicit. Few applications store
large files in the temporary directory, so users may be surprised by
huge precompiled preambles in there.

The new option is always visible in the UI, even if the used libclang
version does not support it. I don't see a clean way to hide the option
if unsupported, because only kdevclangsupport plugin depends on libclang
and can query CINDEX_VERSION_MINOR; KDevPlatformShell manages Language
Support configuration UI and does not depend on the plugin.