Make Detailed completion:Never option work
The <default> element is outside of (not nested in) the <entry> element,
and therefore does not work correctly. When the user chooses the "Never"
item in the Detailed completion combobox on the Language Support tab of
the Configure - KDevelop dialog, KConfig believes that the default value
is selected and removes the completionDetail entry from kdeveloprc.
Then KConfigGroup::readEntry() returns its second argument in
CompletionSettings::completionLevel(), which is MinimalWhenAutomatic
("When Invoked Manually" in the UI). And thus the "Never" choice works
just like the "When Invoked Manually" choice.