SourceFormatterSelectionEdit: order styles for the UI
Currently styles are ordered by name in the UI. The resulting order is
confusing because some predefined styles precede and the rest - follow
user-defined styles. Furthermore, the style names are never displayed in
the UI, so the order looks arbitrary and illogical to the user.

Display the more likely to be selected user-defined styles first, then
the predefined ones. Within these two sections, order styles
case-insensitively by caption (captions are displayed in the UI as style
identifiers).

Optimize SourceFormatterStyle::caption() to speed up the sorting
comparator that uses this function.

All callers of addStyleItem() now know whether or not the added style is
user-defined => add a style category parameter to this helper function
and eliminate a redundant isUserDefinedStyle() call.