grepview: read settings from config for dsearch!
Since 42eb6bb62c28923e5f6f4ff0cee3e8e6ee07bd2e, when m_show is false,
GrepDialog's constructor returns early to optimize by not setting up the
never-to-be-displayed UI. The UI is initialized by reading from config
(saved when a GrepDialog is closed). When GrepDialog::historySearch() is
called, there is no need to read settings from config, because
GrepDialog::nextHistory() uses settings from search history.

Running `dsearch! word` in KDevelop's Terminal tool view invokes
GrepViewPlugin::startSearch("word", ".", false), which in turn calls
GrepViewPlugin::showDialog(false, "word", false). During this call to
showDialog(), a GrepDialog dlg with m_show == false is created, then a
search starts using default-initialized settings (except for the pattern
and search paths, which are set in showDialog()). The help text for
dsearch! contains:
    Same as search, but starts the search instantly instead of showing
    the dialog (using previous settings).
This commit initializes settings from config when show is false in
showDialog() to make dsearch! work as promised. An alternative fix is to
use the last settings entry in search history. But the search history
can be cleared and thus is not dependable. The fix does not affect
historySearch(), because showDialog() is not called in that case.