grepview: don't search in empty-URL paths
An empty search path URL is displayed as a slash (file system root dir)
in the Search History combobox. In contrast, GrepFindFilesThread's
FileFinder::findFiles() searches in the current directory of the
KDevelop process when a search path URL is empty. A single empty search
path is already considered invalid by the UI (see e.g. the first early
return in GrepDialog::setSearchLocations()). This commit discards empty
search paths between semicolons as well. Now the presence or absence of
a semicolon at the end of the search path list makes no difference,
which should match a user's expectations.