formattinghelpers: consider removing the rest of prefix
The added tests reveal the bug fixed by this commit. For example, at the
parent commit the first of the added tests fails like this:
FAIL!  : TestFormattingHelpers::testFuzzyMatching(remove-braces-in-prefix-before-comment-in-text) Compared values are not the same
   Actual   (output)        : "/ comment"
   Expected (expectedOutput): "// comment"

The bug is in PrefixMatcher::selectFuzzyInsertionRemovalOrReplacement():
when neither fuzzy *m_prefixFirst is found in the text, nor fuzzy
*m_textFirst is found in the prefix, both first fuzzy characters are
skipped, in effect matching the first slash of the comment with the
brace removed by the formatter. Skipping a fuzzy *m_textFirst not found
in the prefix is unavoidable only if at least one non-whitespace
non-fuzzy character remains in the prefix. Otherwise, there is another
possibility: keep *m_textFirst and remove the rest of the prefix.
selectFuzzyInsertionRemovalOrReplacement() now weighs this possibility
against the others.