Increase wait-for-state timeout to fix GdbTest
The test fails on my system with the following output:
QWARN  : KDevMI::GDB::GdbTest::testRegularExpressionBreakpoint() current state KDevelop::IDebugSession::PausedState waiting for KDevelop::IDebugSession::PausedState
FAIL!  : KDevMI::GDB::GdbTest::testRegularExpressionBreakpoint() Timeout before reaching state 3
   Loc: [kdevelop/plugins/gdb/unittests/test_gdb.cpp(1857)]

Waiting for 25 (but not 20) seconds is sufficient to eliminate the
failure on my system. This commit increases the timeout to 50 seconds,
which is hopefully enough everywhere.

I suspect that the test became slower and failing on my system because
of GDB version upgrade from 12 to 13.