LanguageController: don't assert that filename is nonempty
I have added this assertion in 59270583c2f1f9bbe30cd24c1fa68229651379b3,
because I was unaware that KDevelop supports an empty file name in a
remote URL. The assertion fails when a remote URL that ends with a slash
is opened in Terminal tool view. For example:
    open! 'http://example.org/'

Early-return from MimeTypeCache::languagesForFileName() when its
fileName argument is empty, because this function's definition does not
support that and MimeTypeCache should not return anything for an empty
file name.

CCBUG: 373213