Save window geometry of several dialogs
Each of the dialogs affected by this commit can use extra space.
Remember user's resizing and moving of a dialog, because its customized
size is likely to be preferable to the default size the next time the
dialog is shown.

Cannot use KWindowConfig::saveWindowSize() in place of
QWidget::saveGeometry(), because the former takes a QWindow* parameter,
but QWindow* QWidget::windowHandle() returns nullptr for each dialog
enhanced by this commit.