TestCodeCompletion: clean up config after every test function
TestCodeCompletion::testIgnoreGccBuiltins() configures Compiler="GCC",
but not parserArguments; TestCodeCompletion::testOverrideExecute()
configures parserArguments=<a C++ standard>, but not Compiler.
Therefore, GCC is unintentionally used in testOverrideExecute() and a
custom C++ standard is unintentionally used in testIgnoreGccBuiltins(),
if the other test function has run at least once on the system, because
config files are persistent.