problemreporter: support closing the document being renamed
When a document A is open and modified, a document B is saved-as at A's
URL, DocumentControllerPrivate::changeDocumentUrl() closes the document
that is being renamed (B). In this case a documentClosed signal is
followed by a documentUrlChanged signal for the just-closed document.

In this scenario ProblemReporterPlugin::documentClosed() removes the
still-open document A's visualizer and
ProblemReporterPlugin::documentUrlChanged() stores B's visualizer under
A's URL. This way the wrong visualizer is preserved, which causes wrong
application behavior. If the user then saves A under a different URL
without restarting KDevelop first, the following assertion in
ProblemReporterPlugin::documentUrlChanged() fails:
    Q_ASSERT(it.value()->document() == document->textDocument());

Detect closing of the wrong document in
ProblemReporterPlugin::documentClosed() to prevent removing the wrong
visualizer (A's). Remove the right visualizer (B's) in the following
invocation of ProblemReporterPlugin::documentUrlChanged().