Remove obsolete function IDocument::isTextDocument()
This function was introduced in
f7990424b239758812745dea31df14124e9990f1, then on the same day the same
developer made it unnecessary and removed all its uses in
544ababfa02e2eae4e6f5b4ae970f4cae5dad4e3. Later other developers started
calling this function in place of IDocument::textDocument(), possibly
assuming it is a more efficient equivalent. isTextDocument() became
equivalent to textDocument() in 9277898e5513238a0fd84aded8c95ac152ae40eb
in order to work around crashes in the code that used isTextDocument()
by mistake.

Now isTextDocument() is never used the way it was originally meant to
be. Remove this function and use textDocument() instead.