Add previous-URL parameter to documentUrlChanged signal
Some slots connected to the documentUrlChanged signal need to know the
previous URL of the renamed document. Each of these slots must invent
its own way to determine the previous URL. This is verbose, error-prone
and inefficient. The previous URL is readily available in code that
changes a document's URL. Thus the previous URL can be helpfully sent
along with the signal at little performance and complexity cost.