OpenDocumentSet: handle text document URL changes
With the problem reporter scope "Open Documents" selected, when a text
document is saved under a different URL, the problems in it are not
highlighted and do not appear in Problems tool view until the user
closes, then opens the renamed document again. The problems for the
previous URL of the renamed document remain listed in Problems tool view
until OpenDocumentSet is destroyed (when another problem reporter scope
is selected or KDevelop is restarted).

Call WatchedDocumentSetPrivate::delDocument() and
WatchedDocumentSetPrivate::addDocument() for the previous and current
URL respectively when the URL of a text document changes in order to
highlight and list the problems automatically (after a short time
delay). The signal IDocumentController::documentUrlChanged implies that
either the renamed document is open or another document at its current
URL is open.