ProblemModel: support closing the document being renamed
When a document A is open and modified, a document B is saved-as at A's
URL, DocumentControllerPrivate::changeDocumentUrl() closes the document
that is being renamed (B). In this case a documentClosed signal is
followed by a documentUrlChanged signal for the just-closed document.

In this scenario ProblemModel::closedDocument() sets the current
document URL to an empty string, then ProblemModel::documentUrlChanged()
compares the current document URL (empty) to the previous URL of the
renamed document. The compared URLs are not equal, so
documentUrlChanged() leaves the current document URL unchanged (empty).
The end result is that A is the active document, but problems in it are
not highlighted and do not appear in Problems tool view until the user
activates another document, then reactivates the renamed one (assuming
the default problem reporter scope "Current Document" is selected).

The current document URL is empty only when there are zero open
documents. The signal IDocumentController::documentUrlChanged implies
that either the renamed document is open or another document at its
current URL is open. Therefore, we can call setCurrentDocument() in
ProblemModel::documentUrlChanged() when the current document URL is
empty to highlight and list the problems automatically in this scenario.