backgroundparser: assert that correct tracker is removed
BackgroundParser::documentClosed() and
BackgroundParser::documentUrlChanged() correctly handle the corner case
when DocumentControllerPrivate::changeDocumentUrl() closes the document
that is being renamed. The added assertions verify that the tracker is
not null and that the URL of a managed text document is not out of sync,
even in the aforementioned corner case.